requirements.txt
streamlit==1.31.0
openai==1.55.3
httpx==0.27.2
torch==2.2.0
torchvision==0.17.0
Pillow==10.2.0
